Modification classe representations.Rectangle

+++ app/src/main/java/grp3/projet/miage/paris10/fr/m1c20152016_planinteractifbu/leplan/representations/Rectangle.java
@@ -9,7 +9,7 @@ public class Rectangle {
     private float largeurContour;
     private int couleurContour;
 
-    public Rectangle(int x, int y, int l, int h, int color, int largeurContour, int couleurContour) {
+    public Rectangle(int x, int y, int l, int h, int color, float largeurContour, int couleurContour) {
         bounds = new Rect(x, y, x + l, y + h);
         paint = new Paint();
         paint.setColor(color);
